refactor(gis-sdk): decompose vectorTiles.ts into multi-function modules
Closes #1228

Split vectorTiles.ts (628 lines) into 7 focused modules under
vectorTiles/ directory following decomposition-analyzer output:

- types.ts: type definitions (VTGenerateConfig, VectorTileRow, etc.)
- utils.ts: internal utilities (feature ID, metadata, geometry stats, etc.)
- encodeFlatGeobufFromFeatureCollection.ts: FlatGeobuf encoding
- generateVectorTilesFromFeatureCollection.ts: core tile generation
- generateVectorTilesFromJsonBuffer.ts: JSON buffer entry point
- generateVectorTilesFromFgbBuffer.ts: FlatGeobuf buffer entry point
- index.ts: re-export public API

Pattern: multi-function (decomposition-analyzer)
Validation: naming 0, threshold 0, API preserved
typecheck: pass, tests: 37/37 pass
