+++ packages/gis-sdk/src/vectorTiles/index.ts
@@ -1,0 +1,13 @@
+// Re-export public API from vectorTiles module.
+export type {
+    FeatureCollectionLike,
+    VTGenerateConfig,
+    VTGenerateResult,
+    VTMetadataContext,
+    VectorTileProgress,
+    VectorTileRow,
+} from './types.js';
+export { encodeFlatGeobufFromFeatureCollection } from './encodeFlatGeobufFromFeatureCollection.js';
+export { generateVectorTilesFromFeatureCollection } from './generateVectorTilesFromFeatureCollection.js';
+export { generateVectorTilesFromJsonBuffer } from './generateVectorTilesFromJsonBuffer.js';
+export { generateVectorTilesFromFgbBuffer } from './generateVectorTilesFromFgbBuffer.js';
